Allow the user to specify an alternate etc directory for poudriere.
The intention is to enable sandboxing all configuration and thus all build
output.  This is helpful for use as part of a larger build system that many
users may use simultaneously on a given host.

To support this, add a global poudriere option [-e etcdir], and if specified,
export it as an environment variable to the command script.  Standardize all
references to poudriere's etc directory as $POUDRIERE_ETC.

While I'm here, clean up how poudriere.sh executes command scripts so that
its case statement only handles special cases.