changed_deps: Fix checking for libraries which are symlinked by running test(1) injail
Must run the test(1) -f in the jail as a symlink references an
absolute path inside the jail, which may not exist in the host.
For example:

Host has /usr/lib/libpcap.so.8
Jail has /usr/lib/libpcap.so.7

The symlink in MNT/usr/lib/libpcap.so points to /usr/lib/libpcap.so.7.
The host dereferences this and does not find .so.7, since it has .so.8.