During sanity check, detect and delete packages which have changed PKGNAMES
This avoids a warning and bug in pkg-repo with duplicate package names
which requires manual intervention. This may not be the best solution,
but it is the lowest risk change to get into 3.0.x while there is an
ongoing effort to change PKGNAMES in ports to not conflict with others.