MFH: Fix usage of cpdup -X/-x.
-X only accepts 1 file containing a list of files/dirs to ignore.
Use -x with .cpignore in each subdir to properly only exclude
/usr/src and /poudriere.

Note that this does not change any behavior as it worked by
accident before. /usr/src was ignore since it was null-mounted
and /poudriere was ignored since it was listed 2nd as -X
which is ignored itself.

Reported by:	John Hein