bulk: Don't rebuild unneeded build deps of run deps that are needed to be built
For example when building ports-mgmt/portupgrade it has a run dependency
on lang/ruby19. ruby19's build dependencies are all added to the queue
regardless of whether or not ruby19 needs to build. If it later is found
to not require a rebuild, its build deps may still remain in the queue
and be rebuilt due to incremental rebuild needs. This could result in
perl (via autoconf dep) being rebuilt when only portupgrade was requested to be
rebuilt with -C.

This is purposely dissabled with -a as it is not needed there. It is
largely only needed when specifying a set of ports to build that are a
subset of what is already in the repository. Pkgclean would avoid this
in normal build cases, but for port testing it can easily happen where
ports that are not needed for the build are rebuilt before this change.