Bring in color selection from DragonFly branch
Brighter hues were selected for:
  - Success
  - Fail
  - Error
  - origin

The phase was changed from light magenta to dark gray.  For -v builds,
this enables success and failures to really stand out.

The builder ID colors were rearranged to put brigher pastels on the
first 7 builders.  The second set of seven builders use darker hues
in the same color order as the first set.