Move build summary after sig-handler builder list, stop showing log info
Powerful boxes can have 60 builders or more.  Have the builder summary
come before the builder list just means it scrolls off immediately on
these setups.  To deal with this, move the builder summary to the end,
after the builder list is displayed.

For aesthetic reasons (and debatable utility), remove the display of the
log with the sig-handler.  This line is often greater than 80 characters
and wraps around on consoles.