Add job_msg_debug function and use it on suppressed distfiles message
A recent commit suppressed the informational message on distfiles mounts
because it wasn't the type of information that should be shown even in
verbose mode.  However, it's appropriate for a debug mode, so an new
function, job_msg_debug, was created to handle it.

The message itself was changed to remove the word "Status" for alignment
reasons.  IMO, the message actually makes more sense with "Status"
removed.

New function discussed with bdrewery.