HTML: Fix file:// and direct access for showing the interface.
Now the HTML files are installed into each log dir using hardlinks
(1b5a26f6). The files support both the "hosted" solution of mapping
serving the static /usr/local/share/poudriere/html files and mapping
/data to /usr/local/poudriere/data/logs/bulk, as well as the older
format of serving the log directory directly, or viewing over file://.
The older format is referred to as "inline".

Currently the "inline" setup only supports showing the build. It could
easily be extended to show the JAIL and INDEX as well, but would require
hardlinking all of the HTML files into the root log dir and jail dir.
This is efficient on space, but not if the user copies the directories
elsewhere without copying hardlinks (which rsync doesn't do by default
for example).

This setup was specifically done to allow copying only the build
directory elsewhere for archiving with a build of packages/world/kernel
and viewing with the interface without extra setup required.