Fix computing deps after process reaper added in 6c9701f.
The general operation of rexec->jexecd is 1 process per jail. When it is
completed all children should be killed off. However during computing
deps there are multiple rexec running against 1 jail and jexecd. When
the first process finished it would kill off all children of the 1
jexecd including deps processes that were not owned by the calling
rexec.

Fix this by using a double fork for child execs. The first fork parent
will become the reaper. This fixes the compute deps case by allowing
every rexec to get its own reapear on jexecd. It adds an extra process
for the 1 rexec to 1 jexecd case but it is a trivial issue.

An existing issue overlooked by jexecd is that during compute deps if ^C
is called in poudriere and it kills off all children, jexecd also needs
to kill off all children. There is currently no signal handler for
jexecd to catch SIGTERM and kill all children. Ideally jexecd could be a
reaper to all child repear processes as well. This results in a panic
currently though in src r278320 and needs to be fixed in CURRENT.