build_queue(): Optimize the case of a job just finishing.
This avoids checking all the status of all jobs when a jobid was just
passed back through the pipe as completing. In this case just focus
on cleaning that job up and moving on to another build on that builder.

When a job is done, only check if idle builders can run a job from the
queue. When all builders are busy continue to check for crashed
builders.