Generalize the coprocess handling.
The locks are used for the stop_html_json case to avoid shutting down while a
lock is held. A trap could be used in the actual coprocess to wait for the lock
to be released, but requires much more testing and rework to enable the
required job-control in the shell.

The atexit/cleanup is used since kill_and_wait is used. If the coprocess
is for some reason blocking it will get a SIGKILL eventually. So there
is no guarantee cleanup was done.