Rewrite lock_acquire() to not spam 'sleep 0.1'
This now uses a lockf(1)-inspired helper named 'locked_mkdir'. It uses
lockf(1)-style locking with open(O_EXLOCK) to serialize 'locked_mkdir'
itself for the mkdir(2) call. The act of creating a directory is
considered obtaining the lock since it is not bound to a specific
process and is atomic. If the directory already exists (or is a file)
then kqueue(2) is used to wait to mkdir(2) once it is deleted. A timeout
is added that is shared between the O_EXLOCK lock and kqueue(2) wait.