Always protect(1) poudriere and its coprocesses from OOM killer.
parallel_run processes are excluded mostly due to the overhead of
protecting them. Those generally only are used pre-build so are
not likely to hit OOM.

The reasoning for this is so Poudriere can do all needed cleanup of its
jails and mounts.

Port builds are _not_ protected. Only Poudriere's control scripts.