Level the ports git integration to the same level as jail
Add a new argument -U to allow to to directly speficy the URL where to
fetch the ports from
Deprecate GIT_URL (but still use it in case already defined in
poudriere.conf. Probably here we should not use it at all and just warn
the user).
Add a new GIT_PORTSURL
Add support for git+ssh and git+https method