When running cpdup, always do it from a common directory with relative paths.
A quick benchmark of cloning 3 builders from a root-on-ZFS, ref on
tmpfs, to tmpfs builders, shows 47% time savings.  Overall on a build
with 3 packages there was a 10% time savings.  It is expected that this
strategy will have larger benefits for long builds.  This avoids
touching various namecache and vnode locks on the parent filesystems
that are not interesting to the tmpfs cloning.

Suggested by:	mjg