port_is_needed(): Rework into pkg_is_needed() to compare PKGBASE.
This is because the queue is based on PKGNAMEs, not ORIGINs.  To
support multiple packages per origin we should not be doing comparisons
on origin if it can be avoided.

This also allows removing the port_deps file which could easily be misused and
create multiple-package-per-origin problems.  It also avoids some extra
overhead.