Pass -n to umount to avoid sync(2).
umount -f avoids sync(2), but without -f it defaults to
syncing if -n is not specified.  In these cases we don't
care to sync(2) the mountpoint or the entire system.
This avoids a ton of contention introduced when -f was
removed.

The -f was removed as it causes the entire process tree to
be walked every time, which is needless contention as well
since we can easily avoid processes from running in these
cases.