Using ptsort, fix PRIORITY_BOOST to also boost needed dependencies.
This adds a USE_PTSORT option (default yes) which can be toggled to use
the older method for testing.

Ptsort (https://www.github.com/dag-erling/ptsort) will give every
package in the dependency tree a priority based on the dependencies it
has.  It also allows setting an initial priority for a package, thus we
can say that gcc starts with a priority of 99, and then everything it
needs will have 99 added to its own priority.  The bug has more
information and details in it on how this all works with the queue.

Fixes #387