Unset UID again from ports_env (after make.conf change) and do it later.
It is done later so that more of the build before building the port
can benefit from it.

There is currently a bug with setting UID=${PORTBUILD_UID} in the
make.conf with the ports framework trying to 'su' to do some operations
which this UID= override causes a loop on.  The ports framework needs
to override UID=0 as a make argument for all of its su executions to
work around this.