Add HTML_JSON_UPDATE_INTERVAL (def: 2) to allow overriding JSON update interval.
This is used by the html_json process to update the .json files in the
log directories for the HTML UI.  Updating less often saves on
exec overhead, but makes the HTML UI more stale.  Depending on the
need of the user this value may be anywhere between 1 and disabled.

Allow setting to 0 to disable entirely, which would break the HTML
UI completely.