Rework compute_deps to optimize the -f case (-a is largely not affected).
This changes compute_deps to do multiple passes over the ports tree.
The first pass is to gather make variables (PKGNAME, *_DEPENDS, etc...).
Then that port is added to a depqueue that will create a gatherqueue
entry for every package that has not get been gathered for.  This
processes repeats until no more gather or dep queue work is needed.

After this, another job is ran to create the needed build queue entries
without any distraction of touching the ports.  This phase cannot be
be inlined back into the gather_vars job since it requires knowing the
PKGNAME of each dependency which won't be known until that dependency
is gathered itself.

The old compute_deps was prone to a problem where dependencies like
ports-mgmt/pkg (and countless others) would be looked up redundantly,
wasting a lot of time running make ports executions.  It also could
get into a situation where -f only specific 1 meta port to build and
thus only 1 compute_deps_port job would startup to process all of its
nested dependencies.  This new method would have the first job just find
the direct dependencies of that meta port, and then the next job would
be parallel against all of the found dependencies.  This can be up to
a 60% time savings for such cases.

The -a case is not very impacted by this since it was already optimized
to only visit every port once.  It did do some things redundantly, but
only a ~4% improvement is seen with this refactoring.