Pass -DNOCLEANDEPENDS when cleaning to not clean dependency dirs.
This speeds up 'make clean' quite a lot on large meta ports.  Since we're
never building dependencies, there shouldn't be much in these directories.
It's possible that something like BUILD_DEPENDS nonexistent:... will trigger
another port to extract as well, but we'll capture that with the
rm -rf wkrdirs/* spam, and upcoming support to just remount wrkdirs like we
do for LOCALBASE.