Fix cache_get_pkgname crash in delete_old_pkgs after 64e219856.
It's possible that a stale package in the PACKAGES directory has an
origin that was not requested to build, but its PKGBASE was.  Just
delete the package.  A pkgclean would remove it as well and this ensures
that no bad packages sit around that the ports build may pick up due
to being the same PKGBASE.