mtree: Avoid descending into all directories.
This seems to have no practical impact on what is detected or not
detected since we always use the same ignore list.  Meaning that
I had thought that foo/* would find that new files in foo/ after
the check would cause an error, but that isn't the case.  So
both 'foo' and 'foo/*' achieve the same test results.   However,
This does cut down on fts(3) and readdir(2) lookups for these
directories which lessens overall contention in the system.