Fix gather_port_vars errors with duplicate ports after 64e219856d.
The problem was duplicate ports in the -f or cmdlist were causing a race
in gather_port_vars.  That code assumes that it is only getting a unique
list of origins to lookup.  It's best to keep it that way as there
are assertions that gather_ports_vars_port is only ran once per origin.