clean.sh: Fix SIGINFO causing spurious crashes in build_pkg.
rm(1) will output which file it is cleaning on stdout when receiving a
SIGINFO.  If a port builder calls clean.sh and receives a SIGINFO at the
wrong time, then the rm(1) being ran will spit out its current file to
stdout and the calling pipe from clean_pool() will receive that and
consider it as a package that is being "skipped".  Normally clean.sh is
only supposed to output packages being skipped on stdout.  In this case
it receives something like "cleaning/deps/gnustep-preview-0.8.5_6/ccache-3.3.4_1"
which it cannot convert back into an origin, and causes a silent crash.
This is later picked up as a 'crashed build: clean_pool'.