Support copying a static ccache into the jail with CCACHE_STATIC_PREFIX.
This uses the existing ccache from the host in the build jail.  This is useful
for using ccache+memcached which cannot easily be bootstrapped otherwise.  The
path to the PREFIX where ccache was installed must be used here, and ccache
must have been built statically.  Note also that ccache+memcached will require
network access which is normally disabled.  Separately setting
RESTRICT_NETWORKING=no may be required for non-localhost memcached servers.