timestamp: Fix not properly handling newlines.
When the full data-ready-to-be-read was printed out, it was
improperly adding a newline in for the next data even
before a \r or \n had been encountered.  This was due
to refactoring in fc8da209bb2e6e8e94e63030abaa6d94502209e8
which made newline change between every read/write.