Make jail -k try harder to cleanup any messes left behind.
This removes the requirement that the mounts actually have directories.
It is possible that there are mounts still around but they are
removed somehow.  This can happen if the jail root is a ZFS
dataset but has nullfs mounts inside of it.  ZFS seems to allow
destroying the dataset even if it has mounts inside of it, thus
poudriere will never unmount them.  This was seen in Issue #296
with an MDMFS bug.