image: Use mdconfig for rawdisk
makefs usage was introduce in commit 0e6b30405bd650a1d9b85672e6dd9f36047f2349
While this works great for embedded images, makefs calculate the number of
inodes based on the number of files. This lead to a really small number and
cause a problem for the type 'rawdisk' as this is mostly used for creating
base disk for virtual machines.

Signed-off-by: Emmanuel Vadot <emmanuel.vadot@gandi.net>