Fix order of jail/builder stop hooks on early exit.
The startup order is jail -> builder.  In a normal bulk run the builders are
stopped at the end of parallel_build(), and later the jail is stopped.

In an early exit however, this incorrect ordering in jail_stop() made the jail
stop hook run before the builder stop hook.  This is problematic for builder
hooks which do something like tmpfs-mount into the reference jail and then
nullfs-mount from the reference jail.  In that case the jail stop hook would
fail to umount the reference jail tmpfs.