delete_old_pkg: Fix with DEPENDS_ARGS.
Mapping an existing package with an origin back to what origin and PKGNAME
might be in the queue now is not straight forward because the origin may be in
it with any number of DEPENDS_ARGS/FLAVOR to produce multiple packages.  What
we can do is look for a matching PKGBASE at the known origin.  From here we
take the closet matching PKGBASE as the new PKGNAME.