Fix fatal errors during dependency calc not being caught.
This fixes err() being called from inside of something like
gather_port_vars which just does an exit.  This exit was
not caught because the command was ran from prefix_stderr_quick
which was not passing the exit status through properly.  It
could probably be fixed, but really any err() inside of
the dep calculation code should still set the dep_fatal
flag to ensure it is not lost.