Allow adding origin@flavor into the build list.
This refactors some of the gatherqueue handling to deal with
having origin and origin@flavor in the build queue
without too much extra overhead to deal with it.

It also reworks gather_port_vars_port to no longer take an inqueue
parameter to note whether the port was listed or needs a queue removal.
Now the queue removal is done from gather_port_vars and the rdep is
passed directly from the queue to note whether it was listed.

The rdep to gather_port_vars_port may now be "metadata" which indicates
that we need to lookup the metadata for the main port.  In the case of
queueing both a FLAVOR-specific and the main port the "metadata" rdep
will be switched back to "listed".