poudriere image: Add embedded type
The embedded type is mostly targeted for arm,arm64 and mips boards.
It create an 20M fat16 partition for storing u-boot, loader and possibly
DTB at one point.
As target arch might be different than the host arch, packages are installed
with pkg -r and specifiying the ABI in the env.
This is a BYOU (Bring You Own U-Boot) image, intalling U-Boot is, for now,
beyond the scope.