Fix handling of main-dep-and-specific-listed test case not passing.
The problem was that origin@FLAVOR would queue a origin metadata
lookup, and then later a dependency would want the origin@DEFAULT
flavor.  This would be thrown away since it was already looked up.
For this case we need to fix the origin and continue processing it
once it is found superfluous.

These problems don't exist with bulk -a since we're guaranteed to
visit the main port/FLAVOR before any specific FLAVORS via dependencies
or even from the listed due to the use of the flavor queue.