Warn on bad PKGNAME dependencies and make fatal with BAD_PKGNAME_DEPS_ARE_FATAL.
Consider bad dependency lines on the wrong PKGNAME as fatal.
For example:
   BUILD_DEPENDS=  p5-List-MoreUtils>=0:lang/p5-List-MoreUtils
If this port's PKGNAME were really "List-MoreUtils" then it would
not be recorded into the resulting package.  The next build with
CHECK_CHANGED_DEPS enabled would consider it a "new dependency"
since it is in the the port but not in the package.  This is usually
a warning but can be made fatal instead by enabling this option.