DEPENDS_ARGS: Deal with python ports which force default 3.
This fixes having www/py-yarl queued (default 3) which depends
on www/py-multidict (default 3) and tries to apply the DEPENDS_ARGS
hack to it only to find it is superfluous.  That would result
in a duplicate PKGNAME failure.  Deal with this kind of
situation more permissively to remove the need to hardcode every
new case of this.

This also allows removing some specific python origin hacks.