jail -u: Verify TORELEASE before performing jail upgrade
In case we use FreeBSD dists and TORELEASE is present, check if it's a release
branch. (verify it even before trying to mount the jail prevents accidents, and
"exit 1" ensures no unnecessary calls and traffic to FreeBSD's update servers
will be performed)

add *-STABLE to the list of possible snapshot versions.

Issue #509