Fix kern.randompid / parallel pid reuse resulting in 100% CPU.
This fixes a 100% CPU problem with kern.randompid set.  Consider if process
A spawns, then process B, then A finishes but before it is removed from
PARALLEL_PIDS with _reap_children, it is spawned again as A and duplicated
into PARALLEL_PIDS.  When _reap_children finally runs it would remove
A, then B, then again try to remove A.  The 2nd removal of A (non-existent now)
would incorrectly duplicate the entire remaining list and basically lead
to a very long `kill -0` loop.

This seems to be a regression from d480361bc2be93183bf1cccb864e5e4be50bbc9c.