Remove bulk/testport -s.
This was a precursor to resuming builds with -B.  It is not
safe to use in general since it skips all incremental build
checks such as changed options, changed PKGNAME, updated
version, changed dependencies, changed default FLAVOR.

-S remains which only skips removing packages if their
dependencies are missing.  This assumes that PORTREVISION
bumps are done properly.  -S can be used for when you
trust PORTREVISION bumps are proper or that the change
made to updated ports are all ABI-safe for downstream
consumers.