pkg_get_*: Use shash to fix handling of empty files.
The old code would check if the file existed and then try
reading from it 100 times.  If that failed then it simply
returned empty data or crashed.  The new version is more
resilient.  If a blank is found then shash_get fails and
the real lookup is done.  Blank files could happen due
to SIGINT.