Add support for zfs image types, including send streams
Adds the ability to create a ZFS disk image in the same style as bsdinstall(8).
The default is to create a full GPT disk image, but is also supports just the raw pool.

Also adds `zfs send` streams as an output format, includes support for sending the
entire pool, just the boot environment (for upgrades), or both.
Supports specifying both zfs+send targets (+full and +be) in a single run.

Extends the existing -i support, to be able to modify existing images.

Adds the -R flag to allow user to specify ZFS send feature flags

Sponsored by: Modirum MDPay
Sponsored by: Klara Inc.