image: adds zsnapshot imagetype
The purpose of this format is to be able to run jails off a zfs snapshot
a couple of files would be generated:
  - NAME-VERSION.manifest.json
  - NAME-VERSION.full.img.gz
  - NAME-latest.manifest.json -> NAME-VERSION.manifest.json (symlink)
and optionaly (if image is generated off a previous snapshot):
  - NAME-VERSION.incr.img.gz
  - NAME-VERSION.modified.files

The full image is the result of zfs send ran off the image builder,
the incremental zfs image is a diff from previous versoin.

No kernel is present on those images, and the purpose is for servers
to download those files by however means (http, ftp, scp, ...),
verify the files hash, and import them on their local zpool.