jails, git: remote may have updated branch non-fastfoward
poudriere ports is expected to be the same as remote. If remote has
updated the objects in non-fast-forward mode (rebase, git push force, ...).
The pull will try to merge objects and will leave dirty merge markers.

This commits removes this behavior and tries to make the local clone
build the exact same objects as those in the remote repository.

Signed-off-by: Arthur Gautier <baloo@gandi.net>