jail -cu: Stop modifying /etc/login.conf in the jail.
This moves the handling to bulk jail startup.

A special case is still done in jail -u for running freebsd-update
(anywhere injail is used).  An earlier commit tried to use
freebsd-update -b but was reverted in commit 54befd6e3976aa58845
as it was apparently modifying files in the host.  Antoine suggested
checking bugzilla for known freebsd-update bugs regarding this.

This now avoids modifying 'jail -m null' jails.