testport.sh: Fix make.conf not being respected
Currently, testport (-c) does not honour make.confs such as can be set
in poudriere.d/{*-}make.conf files, because the __MAKE_CONF variable [1]
is set to /dev/null, introduced in a43fe7cb to fix #545.

Accordingly, when using testport -c, when dialog4ports is run, it presents
information that does not match what is actually used/built because the
expected make environment(s) configurations are not used.

For example, this issue was identified while testing the www/py-autobahn
port, and make config (dialog4ports) was incorrectly displaying
py36-autobahn and py36 specific OPTIONS, when testport was in fact
building py27-autobahn and py27 specific OPTIONS (correctly) via -zpy27,
which contained a DEFAULT_VERSIONS=python=2.7 override.

The issue is not limited to package names and OPTIONS, but anything that
needs to take effect or be modified in the config target.

Conversely, the options command introduced a fix to "Fix make.conf not
being respected in 2d9b712b, introducing a setup_makeconf() function in
common.sh that constructs a make.conf based on global, tree, and set
specific make.conf's, if they are set.

This change utilises the setup_makeconf() function in common.sh, and
applies it to testport.sh to correctly respect make.confs when -c is used.

[1] man make.conf