IGNORE: Restore old accounting method for skipped/ignored.
Because skipped is both pre- and post- build now it needs to still
be accounted for in the queued stat. There are also backwards-compat
problems with the html interface. Just revert to the old
accounting method to resolve the issue.

Revert "Tobuild should account for ignored/skipped being trimmed out of queued"

This reverts commit 383f6f2ab47df89cb746e132c4ba891af1d435cb.

Revert "IGNORE: Don't count trimmed ignored/skipped ports as part of the queued count."

This reverts commit 1fd1aa73ef592d384cae1a5fd006f915507daabc.