Fix hang in exit with SIGINT.
If SIGINT comes in during exit, after it is blocked, while stop_builders
is running it can hang in parallel_run. The children unblock SIGINT
because for most others cases it makes sense to nuke all children when
SIGINT comes in. For cleaning up though we want the cleanup child to
succeed so it does't leave leftover mounts.

Easily seen with 'make -j 25 check' and sending SIGINT as soon as a bulk
test finishes.