install_html_files: Fix build dir not being present.
If we didn't copy the base files into .html then the directory may not
even exist for dest. The problem is another build is locking on the
base dir but also on our build-specific directory. So the lock needs
to be dropped after updating .html.

A shared lock would be better for the relock.