Clean up the svn/git repository after updating a jail to a newer version.
Updating a jail to a newer version grows the .svn / .git metadata folder.
Compact it after updating and before taking a new @clean snapshot to
keep the size down.

It is unlikely that people will modify these checkouts by themselves. A
patch can be used for this when using svn and git supoprts feature branches
for this.

(WIP) Perhaps this needs a new knob in poudriere.conf.
(WIP) Perhaps we want to be more aggressive in the cleanup method.