locks don't need critical sections now
locked_mkdir is the implementation of the lock and supports checking for
stale locks so, in general, it's fine if a process dies while holding it.

Commit 1e49c25ec4e2f4597bec6c16ebdccba984fd239b added this feature
because the html_json process was exiting while holding a lock. We
still need to ensure a consistent stats writing from there though.

Also remove stale comemnt.