image: Add pkgbase support
When building a jail with -m src or using svn/git/etc ... if the user
adds -B it will make packages after buildworld/installworld.
The packages will be placed in ${POUDRIERE_DATA}/image/${JAILNAME}-repo.
When building a poudriere image the user can specify which pkgbase packages
will be installed in the resulting image with -P

Sponsored-by: The FreeBSD Foundation
Signed-off-by: Emmanuel Vadot <manu@FreeBSD.org>