Add support for fetching from artifact.ci.freebsd.org
This adds a new method -m freebsdci which fetches the latest tested sets
from the CI artifacts.
Sometimes head or stable doesn't compile and we waste time and power starting
building a jail to see later that it fails.
This also make it possible to create an armv7 jail from latest tested head which
isn't possible using the ftp method as re@ doesn't publish sets.

Signed-off-by: Emmanuel Vadot <manu@FreeBSD.org>