jail: Add archive subcommand and create method
The archive subcommand (-A) creates a tar file containing the contents
of a jail and it's metadata.  The create subcommand (-c) gains a new
method to create a new jail from an archive, restoring its metadata.
This can be used to save a copy of a jail before updates or to transfer
a jail to another machine.