jail: pkgbase: Add update-packages support
Now that head support update-packages add support in jail update.
Since pkgbase is present in 11/12/13 but update-packages is only in head
and will not be mfc add a fallback to make packages if update-packages rule
doesn't exists.

Also add hooks for pkgbase package build

While here do not allow building packages if not kernel was provided
as make packages expect kernel to be built.