setup_xdev: only set AS if as(1) is present
Notably, as(1) has been removed from base in advance of FreeBSD 13. All
ports that require it now have a build dependency on it. We could just
remove this entirely and force it to use binutils on FreeBSD 11 and 12, but
this is more consistent with how these dependencies are setup for native
builds.

This fixes the build of lang/ocaml, and likely others, on -CURRENT
cross-builds with native-xtools.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>