CCACHE_DIR_NON_ROOT_SAFE: Chmod /root to n55 to allow non-root to access
Also add CCACHE_GROUP and CCACHE_GID to auto add the portbuild user into
the group for configurationg following
https://ccache.dev/manual/3.7.11.html#_sharing_a_cache.
The value default to PORTBUILD_GROUP/PORTBUILD_GID respectively.

The following commands can be ran from the host to set up the directory
for shared writes from the jail's portbuild user.

	pw groupadd portbuild -g 65532
	pw useradd portbuild -u 65532 -g portbuild -d /nonexistent -s /usr/sbin/nologin
	pw groupmod -n portbuild -m root
	CCACHE_DIR=/root/.ccache
	echo "umask = 0002" >> ${CCACHE_DIR}/ccache.conf
	find ${CCACHE_DIR}/ -type d -exec chmod 2775 {} +
	find ${CCACHE_DIR}/ -type f -exec chmod 0664 {} +
	chown -R :portbuild ${CCACHE_DIR}/
	chmod 1777 ${CCACHE_DIR}/tmp

CCACHE_GROUP/CCACHE_GID are not used in the example but if a separate group
were wanted:
	pw groupadd ccache -g 65531
	pw groupmod -n cacche -m root
	chown -R :ccache ${CCACHE_DIR}/
	# poudriere.conf
	CCACHE_GROUP=ccache
	CCACHE_GID=65531

Fixes #832