bulk/testport: Detect *new but disabled* OPTIONS
It is possible that a port maintainer adds a new OPTION that
is disabled by default but changes the behavior of the port
compared to the previous version such that it disables a feature
that was otherwise enabled by default.

While here fix the old pretty-print-config space error and add a new
caching format.

Fixes #900