bulk/testport: Remove -S
This flag was a hack to avoid recursively rebuilding packages but a
fatal flaw is that Pkg records the versions of dependencies. If that
specific version is not available then Pkg breaks. The intent of this
feature was to respect version bumps (and other rebuild checks) rather
than forcing a rebuild of everything. That feature is coming in
Issue #822.

If for some reason this is needed then SKIP_RECURSIVE_REBUILD=1 can be
used in poudriere.conf.

Issue #806