bulk: Improve incremental rebuild by not auto chasing dep updates.
This will now only rebuild a package if `pkg upgrade` would [re]install
it.

This also fixes fetching packages and then deleting them due to
incremental decisions [1].

- This uses pkg's PKG_NO_VERSION_FOR_DEPS to avoid writing out specific
  versioned dependencies in the packages.  This combined with the bulk
  -S behavior of skipping recursive delete is sufficient for proper
  packages.  The recursive delete was originally put in because of the
  deps problems (for `pkg add`) [1] but also just uncertainty.

- This (and `pkg upgrade` already) relies on PORTREVISION chasing when a
  dependency update requires rebuilding consumers.  If a shared library
  dependency soversion is updated then Poudriere _will_ rebuild and `pkg
  upgrade` will reinstall as well.

- This sets PKG_NO_VERSION_FOR_DEPS to no by default.

- Setting PKG_NO_VERSION_FOR_DEPS to yes implies SKIP_RECURSIVE_REBUILD=1.

- We will still delete packages that have versioned dependencies but once
  they are converted they will not be auto deleted recursively again.

- When fetching from remote, with bulk -b, we will look for packages that
  have no versions on their dependencies OR having dependencies matching
  the same versions our local port has.

[1] Consider if Poudriere never did the dependency rebuild chasing itself
and we stored versioned dependencies.

Suppose we have installed 2 packages, consumer and depfoo, both at version
1.0.  Before this change consumer would have a hardcoded dependency of
depfoo-1.0 in its package metadata. When it is installed with `pkg add`, or
`pkg upgrade`, it would require that specific version to be installed.  So if
depfoo is updated to 2.0 and consumer is not rebuilt then consumer would fail
when `pkg check -d` is ran as its dependency is missing.  With this
change `pkg check -d` would pass.

Suppose we have neither pacakge installed. In the remote, Poudriere had
both at version 1.0 but an update to depfoo 2.0 caused it to be rebuilt.
Consumer would not be rebuilt in this example, but would still have a
versioned dependency on depfoo-1.0.  If a user then did `pkg install consumer`
it would error when it attempted to pull down depfoo-1.0 as it no longer
existed and is at depfoo-2.0 now.  With this change depfoo-2.0 would be
pulled down and installed.

At the time of Poudriere's dep chasing, pkg did not auto reinstall much
at all.  Now in pkg_jobs_need_upgrade() it will do reinstalls for things
like changed shlib provides/requires, changed options, changed list of
dependencies, changed conflicts (not handled by Poudriere), changed
provides/requires (not handled by ports nor Poudriere), and changed ABI.
Poudriere will still rebuild in these cases as well so that Pkg can
reinstall the needed fixed package.  Getting to reproducible packages
is a challenge because many of these checks are for user convenience. If
they change a port OPTION they need it to reinstall without them having
to touch PORTREVISION since that field is not for users but is for port
maintainers.  Some other checks like changed deps/requires are
convenience for port maintainers to not have to do a PORTREVISION chase.

Fixes #822
Fixes #954 [1]