bulk [-a] -f: Delete packages unknown/unqueued like pkgclean would.
Unqueued deletion is not done if `bulk -t`, or `tesptport`, is used
or if $DELETE_UNQUEUED_PACKAGES is not set to "yes".  The var can be
set to "always" to allow it during test modes.  This var is not
documented because if someone reports an issue with this feature then
it should be fixed to work for that case or not be enabled by
default; the var would be a temporary workaround.

Unknown files (for example, *.txz) will be deleted unless
$DELETE_UNKNOWN_FILES is not set to "yes".  This var is not
documented because it is expected to be rarely needed and
I don't like the idea of confusing users with more options.

We need a poudriere.conf.sample.full made to track these.

Fixes #493
Issue #915