distclean: Fix deleting from a corrupted list
This could have deleted valid files.

Using awk(1) to dump the data was causing it to buffer up to the disk's
blocksize.  For tmpfs this is 4096 which for a port like
multimedia/gstreamer1-plugins-rust was calling write(2) 5 times
resulting in an intermixed corrupted list of files.  By appending
from inside of awk it ensures it flushes after each record.  We could
just fflush from in there, or avoid awk, or use stdbuf(1) as well.