Delete packages with missing dep PKGBASES
This is dealing with dependencies that have a changed PKGBASE from what the
package is looking for.

A specific example here is https://github.com/freebsd/freebsd-ports/commit/91966715bcf3989bd7c96e5910f0a83952eecdc6
which changed the PKGBASE for net/openldap24-client from openldap-client
to openldap24-client.  Packages such as php74-ldap were depending on
this metadata:
	"openldap-client":{"origin":"net/openldap24-client"}
That would cause pkg-add (during make depends) to try installing
openldap-client-*.pkg.  That would not be found since the openldap package
was now named openldap24-client-VER.pkg, having been rebuilt for a changed
PKGNAME itself.  But reverse deps didn't have the logic to rebuild on
changed dep pkgnames.

Note that this was discovered with the PKG_NO_VERSION_FOR_DEPS feature
because but the root cause is the removal of the recursive deleting of
packages; default SKIP_RECURSIVE_REBUILD==1 causes this.

Issue #822